test-requirements.txt
setuptools>=44.0.0
pytest>=5.4.1
pytest-cov>=2.8.1
pylint>=2.4.4
pylint-exit>=1.1.0
flake8>=3.7.9
anybadge>=1.7.0
requests_mock>=1.7.0
nose>=1.3.7